pyproject.toml
[tool.poetry]
name = "owtask"
version = "0.1.0"
description = ""
authors = ["ellis <ellispridgeon1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9.8,<4.0"
langchain = "^0.1.1"
langchain-experimental = "^0.0.49"
pandas = "^2.2.0"
openai = "^1.8.0"
networkx = "^3.2.1"
tabulate = "^0.9.0"
langchain-community = "^0.0.13"
streamlit = "^1.30.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
